refactor: put components in subfolders
