+++ project-canvas/src/components/StoryMapView/Components/Zoom.tsx
