+++ project-canvas/src/components/StoryMapView/helpers/utils.ts
@@ -1,5 +1,5 @@
 import produce from "immer"
-import { Case, SubActionLevel } from "../types"
+import { Case, SubActionLevel } from "../Types"
 
 export const getRndInteger = (min = 0, max = 100000) =>
   Math.floor(Math.random() * (max - min)) + min
